Token updates
Passing token to the backend. prevent actors accessing the api without a token.

+++ src/bad-each-way-finder-api/bad-each-way-finder-api/Controllers/AccountController.cs
@@ -2,21 +2,31 @@
 using bad_each_way_finder_api_domain.DomainModel;
 using bad_each_way_finder_api_domain.DTO;
 using Microsoft.AspNetCore.Mvc;
+using NuGet.Common;
 
 namespace bad_each_way_finder_api.Controllers
 {
+   
     [Route("api/[controller]")]
     [ApiController]
     public class AccountController : ControllerBase
     {
         private readonly IAccountService _accountService;
-        public AccountController(IAccountService accountService)
+        private readonly ITokenService _tokenService;
+
+        public AccountController(IAccountService accountService, ITokenService tokenService)
         {
             _accountService = accountService;
+            _tokenService = tokenService;
+
         }
-        [HttpGet("GetAccountPropositions/{userName}")]
-        public IActionResult GetAccountPropositions(string userName)
+        [HttpGet("GetAccountPropositions/{userName}/{token}")]
+        public IActionResult GetAccountPropositions(string userName, string token)
         {
+            if (!_tokenService.ValidateToken(token))
+            {
+                return BadRequest("Invalid Token");
+            }
             return Ok(_accountService.GetAccountPropositions(userName));
         }
 
@@ -26,7 +36,12 @@ namespace bad_each_way_finder_api.Controllers
         {
             if (raisedPropositionDto == null)
             {
-                return BadRequest("");
+                return BadRequest("Data is null");
+            }
+
+            if (!_tokenService.ValidateToken(raisedPropositionDto.Token))
+            {
+                return BadRequest("Invalid Token");
             }
 
             var accountPropositions = _accountService.SaveAndGetAccountPropositions(raisedPropositionDto);
@@ -40,14 +55,17 @@ namespace bad_each_way_finder_api.Controllers
         {
             if (raisedPropositionDto == null)
             {
-                return BadRequest("");
+                return BadRequest("Data is null");
+            }
+
+            if (!_tokenService.ValidateToken(raisedPropositionDto.Token))
+            {
+                return BadRequest("Invalid Token");
             }
 
             var accountPropositions = _accountService.DeleteAndGetAccountPropositions(raisedPropositionDto);
 
             return Ok(accountPropositions);
         }
-
-
     }
 }
